requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.3
torch==2.2.1
diffusers==0.26.3
transformers==4.38.1
accelerate==0.27.2
opencv-python==4.9.0.80
huggingface_hub==0.20.3
safetensors==0.4.2
tqdm==4.66.2
scipy==1.12.0
requests==2.31.0